Add wind chill calculator

+++ app/models/Reading.java
@@ -11,10 +11,10 @@ public class Reading extends Model
     public int code;
     public double temperature;
     public double windSpeed;
-    public int windDirection;
+    public double windDirection;
     public int pressure;
 
-    public Reading(int code, double temperature, double windSpeed, int windDirection, int pressure){
+    public Reading(int code, double temperature, double windSpeed, double windDirection, int pressure){
         this.code = code;
         this.temperature = temperature;
         this.windSpeed = windSpeed;
